+++ src/tests/containers/CreateEventPage.test.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { shallow, mount } from 'enzyme'
-import { EventForm } from '../../containers/EventForm'
+import { CreateEventPage } from '../../containers/CreateEventPage'
 import { StaticRouter } from 'react-router'
 
-describe('EventForm', () => {
+describe('CreateEventPage', () => {
   let wrapper
   const props = {
     location: { pathname: '/events/new' },
@@ -13,10 +13,11 @@ describe('EventForm', () => {
     projects: [],
     fetchActiveProjects: jest.fn(),
     loggedInUser: {},
-    createEvent: jest.fn()
+    createEvent: jest.fn(),
+    handleStartDateChange: jest.fn()
   }
   beforeEach(() => {
-    wrapper = shallow(<EventForm {...props} />)
+    wrapper = shallow(<CreateEventPage {...props} />)
   })
 
   it('setsLastLocation with path', () => {
@@ -31,14 +32,14 @@ describe('EventForm', () => {
   it('sets state with projects if received from props', () => {
     const expected = [{ id: 3, name: 'Project1' }]
     props.projects = expected
-    wrapper = shallow(<EventForm {...props} />)
+    wrapper = shallow(<CreateEventPage {...props} />)
     expect(wrapper.state().projects).toEqual(expected)
   })
 
   it('sets state when props are updated', () => {
     const expected = [{ id: 3, name: 'Project1' }]
     wrapper.setProps({ projects: expected })
-    wrapper = shallow(<EventForm {...props} />)
+    wrapper = shallow(<CreateEventPage {...props} />)
     expect(wrapper.state().projects).toEqual(expected)
   })
 
@@ -46,7 +47,7 @@ describe('EventForm', () => {
     const context = {}
     wrapper = mount(
       <StaticRouter context={context}>
-        <EventForm {...props} />
+        <CreateEventPage {...props} />
       </StaticRouter>
     )
 
@@ -62,38 +63,5 @@ describe('EventForm', () => {
     submitForm.simulate('submit')
 
     expect(props.createEvent).toHaveBeenCalledTimes(1)
-  })
-
-  it('sets state when startDate is changed', () => {
-    const startDateSelect = wrapper.find('DatePicker').filterWhere(item => {
-      return item.prop('name') === 'startDate'
-    })
-
-    startDateSelect.simulate('change', new Date('04/05/2019'))
-
-    expect(wrapper.state().startDate).toEqual(new Date('04/05/2019'))
-  })
-
-  it('sets state when endDate is changed', () => {
-    const context = {}
-    wrapper = mount(
-      <StaticRouter context={context}>
-        <EventForm {...props} />
-      </StaticRouter>
-    )
-    const eventForm = wrapper.find('EventForm')
-    eventForm.setState({ repeats: 'weekly', repeatEnds: 'on' })
-
-    const weekdaysSelect = wrapper.find('Select').filterWhere(item => {
-      return item.prop('name') === 'weekdays'
-    })
-
-    weekdaysSelect.simulate('change', ['Monday'])
-    const endDateSelect = wrapper.find('DatePicker').filterWhere(item => {
-      return item.prop('name') === 'endDate'
-    })
-    endDateSelect.props().onChange(new Date('2019-05-26'))
-
-    expect(eventForm.state().endDate).toEqual(new Date('2019-05-26'))
   })
 })
